+++ ModelSwapperSkins/BoneMapping/BoneInitializer.cs
@@ -57,6 +57,7 @@ namespace ModelSwapperSkins.BoneMapping
             AddCustomBoneInitializerRules(BoneInitializerRules_Lemurian.Instance);
             AddCustomBoneInitializerRules(BoneInitializerRules_LemurianBruiser.Instance);
             AddCustomBoneInitializerRules(BoneInitializerRules_LunarExploder.Instance);
+            AddCustomBoneInitializerRules(BoneInitializerRules_Merc.Instance);
             AddCustomBoneInitializerRules(BoneInitializerRules_Mage.Instance);
             AddCustomBoneInitializerRules(BoneInitializerRules_Parent.Instance);
             AddCustomBoneInitializerRules(BoneInitializerRules_Railgunner.Instance);