+++ ModelSwapperSkins/BoneMapping/InitializerRules/BoneInitializerRules_Merc.cs
@@ -1,0 +1,32 @@
+﻿using RoR2;
+using UnityEngine;
+
+namespace ModelSwapperSkins.BoneMapping.InitializerRules
+{
+    public sealed class BoneInitializerRules_Merc : BoneInitializerRules_AutoName
+    {
+        public static new BoneInitializerRules_Merc Instance { get; } = new BoneInitializerRules_Merc();
+
+        public override bool AppliesTo(BodyIndex bodyIndex)
+        {
+            return bodyIndex == BodyCatalog.FindBodyIndex("MercBody");
+        }
+
+        protected override BoneInfo getBoneInfo(Transform modelTransform, Transform potentialBoneTransform)
+        {
+            BoneInfo bone = base.getBoneInfo(modelTransform, potentialBoneTransform);
+
+            switch (bone.Type)
+            {
+                case BoneType.Toe1L:
+                    bone.RotationOffset *= Quaternion.Euler(0f, 90f, 0f);
+                    break;
+                case BoneType.Toe1R:
+                    bone.RotationOffset *= Quaternion.Euler(0f, 270f, 0f);
+                    break;
+            }
+
+            return bone;
+        }
+    }
+}